IPC-101: Move Timestamp to own module.
