+++ ipld/resolver/src/behaviour/membership.rs
@@ -26,8 +26,8 @@ use tokio::time::{Instant, Interval};
 
 use crate::hash::blake2b_256;
 use crate::provider_cache::{ProviderDelta, SubnetProviderCache};
-use crate::provider_record::{ProviderRecord, SignedProviderRecord, Timestamp};
-use crate::stats;
+use crate::provider_record::{ProviderRecord, SignedProviderRecord};
+use crate::{stats, Timestamp};
 
 use super::NetworkConfig;
