+++ ipld/resolver/src/lib.rs
@@ -9,6 +9,7 @@ mod provider_record;
 mod service;
 mod signed_record;
 mod stats;
+mod timestamp;
 mod vote;
 
 #[cfg(any(test, feature = "arb"))]
@@ -20,3 +21,4 @@ pub mod missing_blocks;
 pub use behaviour::{ContentConfig, DiscoveryConfig, MembershipConfig, NetworkConfig};
 pub use client::Client;
 pub use service::{Config, ConnectionConfig, NoKnownPeers, Service};
+pub use timestamp::Timestamp;
